Board Setup
Created a board node that dynamically draws lines based on the number of squares on each file and rank.
Created the game room scene and placed the board node in it. The board node is scaled to match the window size.
Created a board custom resource which will be used to save and load board types.
Imported some assets for a basic kanic shogi piece set, as well as a wood texture for the board and some UI elements for the promotion window and square highlights.